+++ src/events/DoxterEvent.php
@@ -1,0 +1,8 @@
+<?php
+namespace selvinortiz\doxter\events;
+
+use yii\base\Event;
+
+class DoxterEvent extends Event {
+    public $source;
+}